Added in TODO/FIXMEs in app.py, started working on refactoring code in the backend

+++ backend/setup.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
 # TODO: Write whatever else we need in the future to populate the database
-rm ../database/everything.db
-python3 init.py
-python3 utilFunc.py
-python3 utilFuncNotes.py
-python3 utilFuncChat.py
+if [ -f "../database/everything.db" ] ; then
+    rm ../database/everything.db
+fi
+python3 app.py
